Fix type check, and add asserts submodule

+++ raw_socket_listener/listener.go
@@ -441,7 +441,7 @@ func (t *Listener) readPcap() {
 					}
 
 					// Invalid length
-					if ihl * 4 > ipLength {
+					if int(ihl * 4) > ipLength {
 						continue
 					}
